requirements.txt
streamlit==1.31.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
ultralytics>=8.0.0
diffusers>=0.24.0
transformers>=4.36.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
pyyaml>=6.0.0